requirements.txt
# Web framework
Django==4.2
pillow

# API Package
djangorestframework==3.15.2
django-filter==24.3
djangorestframework-simplejwt==5.3.1

# API Documentation
drf-yasg==1.21.7

# Templated Email package
django-mail-templated==2.6.5

# Fake data generator dependency
Faker==28.1.0

# for zarinpal api dealing
requests

# Testing
pytest-django==4.9.0
# Deployment